city-event-crawler/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
geopy>=2.4.0
aiohttp>=3.9.0
apify-client>=1.6.0
praw>=7.7.0
tweepy>=4.14.0
cachetools>=5.3.0
python-dotenv>=1.0.0
selenium>=4.15.0
playwright>=1.40.0
lxml>=4.9.0
feedparser>=6.0.0
